+++ runtime/bin/io_natives.cc
@@ -87,6 +87,7 @@ namespace bin {
   V(NetworkInterface_ListSupported, 0)                                         \
   V(Platform_NumberOfProcessors, 0)                                            \
   V(Platform_OperatingSystem, 0)                                               \
+  V(Platform_OperatingSystemVersion, 0)                                        \
   V(Platform_PathSeparator, 0)                                                 \
   V(Platform_LocalHostname, 0)                                                 \
   V(Platform_ExecutableName, 0)                                                \