+++ CHANGELOG.md
@@ -26,6 +26,8 @@
     will be thrown. Note: if the duration specified in 'timeout' is greater than
     the system level timeout duration, a timeout may occur sooner than specified
     in 'timeout'.
+  * Added `Platform.operatingSystemVersion` that gives a platform-specific
+    String describing the version of the operating system.
 
 * `dart:core`
   * The `Uri` class now correctly handles paths while running on Node.js on